+++ packages/core/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./redux";
 export * from "./api";
 export * from "./RoomParticipant";
-export * from "./version";
 export { createServices } from "./services";
+export * from "./utils";