+++ packages/core/src/redux/slices/signalConnection/index.ts
@@ -1,10 +1,11 @@
-import { createSlice, ThunkDispatch, AnyAction, PayloadAction, createSelector } from "@reduxjs/toolkit";
+import { createSlice, ThunkDispatch, PayloadAction, createSelector, UnknownAction } from "@reduxjs/toolkit";
 import { RootState } from "../../store";
 import { createAppThunk } from "../../thunk";
 import { createReactor, startAppListening } from "../../listenerMiddleware";
 import { selectDeviceCredentialsRaw } from "../deviceCredentials";
 
 import {
+    AudioEnableRequestedEvent,
     AudioEnabledEvent,
     ChatMessage,
     ClientKickedEvent,
@@ -17,6 +18,7 @@ import {
     NewClientEvent,
     RoomJoinedEvent,
     RoomKnockedEvent,
+    RoomLockedEvent,
     RoomSessionEndedEvent,
     ScreenshareStartedEvent,
     ScreenshareStoppedEvent,
@@ -27,19 +29,23 @@ import { Credentials } from "../../../api";
 import { appLeft, selectAppWantsToJoin } from "../app";
 import { signalEvents } from "./actions";
 
-function forwardSocketEvents(socket: ServerSocket, dispatch: ThunkDispatch<RootState, unknown, AnyAction>) {
+function forwardSocketEvents(socket: ServerSocket, dispatch: ThunkDispatch<RootState, unknown, UnknownAction>) {
     socket.on("room_joined", (payload: RoomJoinedEvent) => dispatch(signalEvents.roomJoined(payload)));
     socket.on("new_client", (payload: NewClientEvent) => dispatch(signalEvents.newClient(payload)));
     socket.on("client_left", (payload: ClientLeftEvent) => dispatch(signalEvents.clientLeft(payload)));
     socket.on("client_kicked", (payload: ClientKickedEvent) => dispatch(signalEvents.clientKicked(payload)));
     socket.on("audio_enabled", (payload: AudioEnabledEvent) => dispatch(signalEvents.audioEnabled(payload)));
     socket.on("video_enabled", (payload: VideoEnabledEvent) => dispatch(signalEvents.videoEnabled(payload)));
+    socket.on("audio_enable_requested", (payload: AudioEnableRequestedEvent) =>
+        dispatch(signalEvents.audioEnableRequested(payload)),
+    );
     socket.on("client_metadata_received", (payload: ClientMetadataReceivedEvent) =>
         dispatch(signalEvents.clientMetadataReceived(payload)),
     );
     socket.on("chat_message", (payload: ChatMessage) => dispatch(signalEvents.chatMessage(payload)));
     socket.on("disconnect", () => dispatch(signalEvents.disconnect()));
     socket.on("room_knocked", (payload: RoomKnockedEvent) => dispatch(signalEvents.roomKnocked(payload)));
+    socket.on("room_locked", (payload: RoomLockedEvent) => dispatch(signalEvents.roomLocked(payload)));
     socket.on("room_session_ended", (payload: RoomSessionEndedEvent) =>
         dispatch(signalEvents.roomSessionEnded(payload)),
     );